company-smart-access-system/services/auth-service/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
pydantic
